Make Arquillian testing work again after pac4j
Adding pac4j broke Arquillian testing, because of another classloader
issue, this time, with HK2.  It seems we need to make a definitive
switch to using a managed container instead of an embedded container.
But for now, we don't use pac4j indirect clients, so do a workaround
that avoids the problem with the HK2 and classloading.