Return an ErrorResult when security or input error means a 400/403
In cases where there is an authentication error, or when the input
data to a POST request is invalid, you get a status code of 400 or
403, but you also get an HTML result that is generated by Tomcat,
because there is otherwise no body.

Now, intercept those cases, and return an ErrorResult (which will be
in either JSON or XML, depending on the request) containing more
useful information.