Refactor ValidationUtils to use nodeModifier parameters better
The duplication of code between methods that either take or don't take
a nodeModifier parameter was not good.

Refactor those methods, so that methods that take a nodeModifier
parameter can accept null for that parameter, and then behave
appropriately if the actual parameter is null (i.e., apply
addPropertyNode(fieldName).addConstraintViolation()).

Then refactor the methods that don't take a nodeModifier parameter so
that they just call the corresponding method that _does_ take such a
parameter, passing in null for it.