Add logstash-logback-encoder; update logback/slf4j to match
Add the logstash-logback-encoder library, and define a logger to use
it.

Update the logback and slf4j libraries to meet the needs of this
library. (When using logstash-logback-encoder, it output a log message
suggesting to upgrade logback. And the newer version of logback has a
dependency on a newer versin of slf4j.)