During test, deploy Sesame to same managed container
Instead of creating an embedded Tomcat in which to deploy Sesame, use
the same Tomcat managed container that we use for the webapp under
test.

Then, we have the problem of making sure that Sesame starts with a
clean working directory each time. So, we put that initialization in a
separate TestNG suite.