Ensure logback isn't shut down before our own contextDestroyed()
With the newer release of logback that we're now using, logback was
(by default) installing its own ServletContextListener which shuts
down logging.

That shutdown was happening before our own
ApplicationContextListener's contextDestroyed() was run. So any
logging done there was getting lost.

So instead, disable logback's installation of its own
ServletContextListener, and defer to it at the end of our own.