Add Java implementation of JayBizzle CrawlerDetect
Add in some of the source files from JayBizzle CrawlerDetect, and an
awk script that converts them into the source of a Java class with a
method for determining if an user agent specified in HTTP headers
corresponds to a known bot/crawler.

Use the class in Logstash logging.

Add a test class.