API methods to get full details of related entities
Add an API method to get the full details of the related entities of a
vocabulary.

Enhance the API method that gets a vocabulary, to allow for inclusion
of the full details of the related entities.

This involves enhancing the registry schema, to allow for returning a
list of related entities, and for returning the full details of a
related entity, embedded within a vocabulary element.