Make registry schema vocabulary id not required
In the registry schema, make the vocabulary id attribute not required.
(When uploading a new vocab, the id is not only not required, it must
not be specified!)  This has the effect of changing the type of the
getID() method from int to Integer. So now we can do a null test
during validation.

Refactor the the CheckNewVocabularyImpl to use a static import for the
interface name; this makes the validation code much tidier and will
help avoid copy/paste errors in future.