Make registry schema related entity id not required
In the registry schema, make the related entity id attribute not
required.  (When uploading a new related entity, the id is not only
not required, it must not be specified!)  This has the effect of
changing the type of the getID() method from int to Integer. So now we
can do a null test during validation.