Support auto-generation of DbUnit DTDs for the databases
Add a standalone class to run generation of DbUnit DTDs for the
databases.  Add an Ant task to invoke it.  Use it to generate an
updated DTD for the Toolkit, and a DTD for the roles database.  Have
the DTD for the registry database generated as part of the
dependencies for testing.