Restructure related-entity-ref element of registry schema
In the registry schema, the representation of related entity
references within a vocabulary was not good.  If the same RE was
related to a vocabulary with multiple relations, there were multiple
RE ref elements for the same RE. And if the full REs are included,
they are duplicated as a consequence.

So now, introduce a <relation> subelement within the
<related-entity-ref> element, to allow the one <related-entity-ref>
element to represent multiple relations to the same RE.