Have .../versionArtefacts/conceptTree return a String on an error
The return type on error (400) of the
/api/resource/versions/{versionId}/versionArtefacts/conceptTree
API method was modelled as an ErrorResult. There is no mapping of that
to the MIME type text/plain, so the 400 result gets turned into a 500
result and the error message doesn't get any further.

So define the return type on error (400) as a String, and don't create
ErrorResult objects, but return a plain String as the error message
instead.