Restructure related-vocabulary-ref element of registry schema
This follows on from commit 2bc284cf3f35c5b5446ce9e072796d7ae460b857.

In the registry schema, the representation of related vocabulary
references within a vocabulary was not good.  If the same RV was
related to a vocabulary with multiple relations, there were multiple
RV ref elements for the same vocabulary.

So now, introduce a <relation> subelement within the
<related-vocabulary-ref> element, to allow the one
<related-vocabulary-ref> element to represent multiple relations to
the same vocabulary.