Add API method to get all related entities, filtered by type
Add an API method to get all related entities.  There is an optional
related entity type parameter to use as a filter.

This method is intended to be used, for example, to provide a list of
suggestions in a dropdown.

For this reason, related entity identifiers are not included in the
results. I.e., once a "selection" has been made, use the
getRelatedEntityById method to get the full details.