Add support for getting top-level details of related vocabularies
Add "vocabulary" as an optional subelement of the
"related-vocabulary-ref" elements in the registry schema.

Turn the includeRelatedEntities parameter of the getVocabularyById API
method into a includeRelatedEntitiesAndVocabularies parameter, that
also inserts the top-level metadata of related vocabularies.

Change the behaviour of getVocabularies so that it _doesn't_ add
related entity references and related vocabulary references.