Add Solr search API method, embedded Solr for automated testing
Add an API method to do Solr searches.

Add an embedded Solr instance for automated testing.

Add support for a Solr "schema version" property, and have the
CreateSchema class set it.

Add a test that ensures that the embedded Solr starts up and has the
schema populated for use by tests.