Change Swagger response type of search API method to Object
The generated JavaScript Client API library doesn't understand the
response from the search API method given the combination of the MIME
reponse type of JSON and the method response type of "string". So,
since the data that comes back from Solr is indeed a JSON object,
change the method response type to "object" to see if that works
better.