Add do-poolparty-harvest flag to versions for registry schema and db
Add a do-poolparty-harvest flag to the specification of version in the
registry schema and the database version-json.

Have the migration process set the flag to true for all versions in
all projects that have a PoolParty ID.

Require that when creating a vocabulary, the flag may only be
specified as true if a PoolParty project is specified.