Add tests for temporal utils; restore logging of test method start
Add tests of the utility methods for temporal database entities.

Fix the implementation of isHistorical() so that it now considers data
to be historical if the end date is exactly now.

In introducting the Arquillian Suite Extension, I had failed to note
that the @BeforeMethod-annotated logTestNameBefore() method was no
longer being run before _every_ test, but only before the dummyTest()
defined in the same class.

It seems to work to move the logTestNameBefore() method up to
ArquillianBaseTest, so I have done that.