Tidy TestNG annotations across the inheritance hierarchy
OK, so I didn't understand the full consequences of breaking up the
test suite into multiple classes.

The behaviour of @BeforeMethod-annotated methods was working
correctly; I didn't appreciate that these _do_ have to be "present" in
every class to which they are intended to apply; and this can be done
using inheritance. So the observed behaviour after the introduction of
Arquillian Suite Extension (i.e., the @BeforeMethod-annotated method
in a subclass not being applied to the other subclasses) was not a
defect. Remove the comment suggesting it might be.

Remove unnecessary "groups" settings within @Test and @BeforeSuite
annotations; the groups settings are determined by inheritance and by
application of the class-level settings at the method level.