Support DbUnit full database comparison; cache IDatabaseConnections
Support DbUnit comparison between the current state of an entire
database and an XML spec of the expected contents.

In order to do this, refactor the DbUnit methods in
ArquillianTestUtils to cache an IDatabaseConnection instance for each
database, and reuse those instances for all tests.

Close these connection instances during server-side shutdown.

The server-side test shutdown (@AfterSuite-annotated method) was not
being run at the _end_ of the suite, but after the first test! So
force server-side shutdown to happen at the end by defining an
explicit Shutdown test in arquillian-tests.xml.