Restore JaCoCo coverage stats collection
JaCoCo coverage statistics collection was found to be broken: it
reported almost zero coverage.

It seems I broke it when switching from the embedded Tomcat container
to the managed Tomcat container.

So we need to specify the JaCoCo agent when starting up the
container's JVM.

Because of the general issue of resolving paths given in configuration
files, the solution now requires the specification of two envrironment
variables for the JaCoCo agent and the JaCoCo output file; those
environment variables must specify absolute paths.

If you use the Ant targets, this is done for you.

If you use Eclipse or another IDE to run tests, you'll have to set
those environment variables yourself.