Add model test: deleting a simple draft vocabulary; update first test
Add another test of the registry model.  The database is populated
with an existing simple vocabulary (with only top-level metadata and
some related entities; no versions) that has only a draft instance.
That draft instance is deleted.

Update the first test (of deleting a current vocabulary) to make it
more "interesting". Also, fix the fact that the input data did not
represent a consistent database.