Ensure ValidationMessage.properties is in the right place during test
ArquillianBaseTest used addAsWebInfResource(p.toFile()) to add the
resources in the WebContent directory into the archive to be tested.
This had the effect of stripping all but the basename of the file, so
that the directory path was lost.

In particular, ValidationMessage.properties ended up in the archive in
/WEB-INF rather than in /WEB-INF/classes.

Now, use a variant of addAsWebInfResource() that allows specifying a
target path, and make sure to preserve the full relative path of the
source file when adding to the archive.