Make version and AP Ids optional in registry schema
We will need to allow the portal (and users) to omit version Ids and
access point Ids when creating vocabularies and doing updates via the
API. So don't make them required in the registry schema.

Adjust the validation of vocabulary data to match the change in the
code generated by xjc.  (The Java type changed from int to Integer;
this allows correctly determining the absence of a value.)