Have DbUnit use JDBC queries with schema set to "PUBLIC"
When working with the roles database, it appeared to DbUnit that the
ROLES table had three extra columns.

It turns out that these three columns are those of H2's
INFORMATION_SCHEMA.ROLES table.  Because we were creating DbUnit
H2Connections by passing in null for the schema, when DbUnit requested
table metadata using JDBC calls, it was getting back metadata across
_all_ schemas.

So now, specify the schema name "PUBLIC" explicitly. (That's the name
used by H2 as the default schema.)