Add API method to get instance of vocabulary for editing
Add an API method to get the instance of a vocabulary that should be
edited.  So, if there is a draft, the draft instance is
returned. Otherwise, the current instance is returned.
In any case, authentication is checked.

Because getting the draft instance relies on the model system,
currently only the top-level metadata and VREs are returned for
drafts.