Fix getCurrentVocabularyRelatedVocabulariesForVocabulary and extend
The existing getCurrentVocabularyRelatedVocabulariesForVocabulary had
a defect: it grouped results by the vocabulary Id of the relating
vocabulary, not the Ids of the related vocabularies.  Now, group
results by the vocabulary Ids of the related vocabularies.

Add a variant of the method that takes an EntityManager parameter.

Add another method that gets draft records.