Don't use full package name in test data directory names
The use of the full package name in the names of directories
containing test data is annoying.

This was all about having the data in a structure that could be used
by an Arquillian extension ... that we ended up not using.

So get rid of the package name from the directory names.

Adjust the test infrastructure accordingly.
Fix up test methods accordingly.