Keep model currentR[EV]sAndRelations maps correct through changes
In both VocabularyRelatedEntitiesModel and
VocabularyRelatedVocabulariesModel, have applyChangesCurrent()'s
visitor class update the currentR[EV]sAndRelations maps with
additions/deletions matching what we do to the database, as is already
done by applyChangesDraft() to draftR[EV]sAndRelations.

This is part of making model methods non-destructive, so that it is
possible, e.g., to read off the updated model after making a change,
without having to fetch everything from the database afresh.