Replace the basic Solr test with a much better one
The testSolrOK() test method was mostly useless, since it didn't help
us diagnose the fact that the schema creation needs to be flushed.

Replace it with a much better test, testSolrIndexing1(), which
actually does an indexing operation and a subsequent query.

To make it work, now do an explicit commit after schema creation.