Add instructions for externalizing logback.xml and other files
Add instructions to context.xml.sample that show how to externalize
files that otherwise must be present in the webapp's classpath.

This is particularly needed for logback.xml if there are going to be
multiple identical WAR files deployed on the same server, since
without taking care of this, the webapps will write to the same log
files.