Add source column to access_points table and add to schema
We need to make the source attribute "first-class", i.e., to have it
as a full database column, rather than hidden in the JSON data column.

We also need to make it visible to API users, including the portal.

So, add it to the database schema, and expose it through the registry
schema.