More work on WorkflowProvider interface and implementations
Add a setEntityManager() method to the interface, with a default,
empty implementation. Add override implementations to a number of
providers.

Add a default implementation for defaultPriority(), so that the
existing implementation in PoolPartyBackupProvider can be removed.

Tidy the comments in the Subtask class a little.

Improve the comments in WorkflowProvider, and flesh out the
deleteVersionArtefact() method.