Validate uploadId attributes of file access points
Create an AuthorizationValidationUtils class.  In that class, define a
method checkAuthorizationForContent() to check that the user is
authorized to "work with" vocabulary content specified in registry
schema format.

For now, that means checking the upload Ids of file access points.

Call the checkAuthorizationForContent() method when creating/updating
vocabularies.

Refactor CheckVocabularyImpl's checking of access points.
Add validation for file access points.