Do file copying and complete info for file access points
When creating a file access point, now do the copying of the file from the upload directory.

Because the target directory is based on the vocabulary and version,
make them available when they're needed, by creating and passing in a
TaskInfo object.

This needed changes to VocabularyModel and VersionsModel so that
AccessPointModel can get the right Vocabulary instance when it needs
to.