More work on subtask generation from model classes; force-workflow
Have the AccessPointModel work harder at deciding when a subtask is
required. To be specific: work out when a (re-)import is required.

Have VersionModel do subtasks based on changes to the version-level
flags of an existing version.

Support a "force-workflow" attribute on the version schema element, to
force workflow processing for an existing version.