Add implied subtasks for ResourceMapTransformProvider
Have VersionModel add subtasks for ResourceMapTransformProvider based
on the doImport and doPublish flags, and on the IMPORT and PUBLISH
subtasks that are going to be done.

Make ResourceMapTransformProvider implement the WorkflowProvider
interface.

Enhance Subtask to store the provider Class object explicitly, not
just as a String.

Add a method to ProviderUtils to get the Class object for a provider.