Migrate IRI lookup resolver and its tests
Migrate the global IRI lookup resolver, and its test cases.

Retire the ResourceNotFoundException class; just generate the Response
in a helper method.

To do this, import the TemporalUtils class in database entity and DAO
classes for entities that are not themselves temporal, but which have
at least one extraQuery that needs TemporalUtils.