Redo RegistryIRILookupTests test data based on ResourceMap tests
The methods of RegistryIRILookupTests promise that "Lookup data is
based on the correct result of" the corresponding
TransformProviderTests.testResourceMapTransformProviderX methods.

On initial migration a few commits ago, I broke that promise.

Now, having migrated the testResourceMapTransformProviderX methods, go
back to the RegistryIRILookupTests test data and make the claims true
again.