Support admin API methods through Swagger; move method to get drafts
Bit the bullet and expose admin methods through Swagger.

Add the necessary bits and pieces to check that a user is a superuser,
and to return a suitable error if not.

Move the functionality of getting all vocabs including drafts out of
the public-level API and make it an admin-level API method.