Don't pass system-generated access points around when updating
Sigh, so many oversights in the handling of access points, leading to
defects in the model and inconsistencies in the API.

For the getVocabularyByIdEdit() method, don't include
system-controlled access points in the return value. This means that,
e.g., the portal doesn't have to worry about ignoring them.

In AccessPointsModel, ignore system-controlled access points within
applyChanges(). This will make it consistent with
VersionArtefactsModel, which seems to show the way I meant to do
applyChanges() in AccessPointsModel.

Don't even _try_ to do workflow processing in applyChangesDraft() when
adding a draft row for an existing access point, i.e., by calling
insertAccessPoint. Just clone the existing access point.

Oh dear, the addition of a version database entity in VersionModel
applyChanges() wasn't visible to AccessPointsModel. Now, make a
note of when that happens, and re-populate sub-models in that case.

Oh dear (again), VersionModel didn't invoke the notifySetNowTime() and
notifySetModifiedBy() methods when (re-)populating sub-models. Now, do
that.