Add final txn.commit() to testJsonTreeTransformProvider1
Oh dear, the testJsonTreeTransformProvider1 test method did not do
txn.commit() on successful completion. It turns out that this had very
bad consequences: a subsequent test (to be committed in a future
commit) that used the version_artefact_id table would fail due to
a concurrent modification exception from H2.

It seems it is essential to close each transaction either with commit
or rollback, in order to ensure independence of tests.