Define equals() and hashCode() for Role
If the user is a "superuser" or "content manager", all organizational
roles are added to their profile. The bulk addition is done using
objects fetched from a separate database query, and those objects were
not "equal" to existing roles that actually do have the same role id,
etc.

It seems that when I implemented this feature in
RolesUtils.updateRoles() I didn't think about adding equals() and
hashCode() methods to the Role class.

So, you could see duplicates in the output of the userData API method.

Oops. Add equals() and hashCode() methods to the Role class.