Add testApplyChangesCurrentVoVREVeAPVAW2 and adjust Json* transforms
Add the testApplyChangesCurrentVoVREVeAPVAW2 test.

Working on this test showed me that I had inadvertently changed the
bheaviour of the JsonListTransformProvider and
JsonTreeTransformProvider transform providers, in the case where there
was no valid RDF input being processed, in which case, these providers
were producing an empty result, but nevertheless generating an output
file and adding a VersionArtefact. Change the behaviour so that the
transform still "succeeds" but _doesn't_ generate an output file and
_doesn't_ add a VersionArtefact. Also, be careful in this case to make
historical any current VersionArtefact.

Patch testApplyChangesCurrentVoVREVeAPVAW1 to cope with this change of
behaviour.