Tweak HTML validation to suit TinyMCE, and apply cleaning
TinyMCE insists on inserting stuff into rel attributes of a tags.

So, tweak the validation to allow any content for rel attributes, on
the basis that we will apply cleaning. Since we are going to do that
now, also allow arbitrary target attributes.

Oops, somehow completely omitted to apply cleaning to HTML metadata.
Do that now, applying our settings for target and rel attributes.