Don't define MySQL triggers
We can't use MySQL triggers as-is, because JPA keeps back some
inserts/updates, and so the sequence of inserts/updates actually sent
to the database may be different from what the code actually does.

Hmm, sigh, we probably didn't notice this problem because they were
only defined for MySQL, not for H2.