Don't fail creating a PoolParty harvest delete, if no PP project
The workflow creates a dummy PoolParty harvest delete, if workflow is
forced and the PoolParty harvest flag is unset or false.

The creation of the Subtask object failed, if no PoolParty project was
specified for the vocabulary.

Make sure that creation of the Subtask succeeds, in this case, i.e.,
specifying a harvest delete.