Be careful when migrating related entities not to create duplicates
Oops, the "canonicalization" process for related entities could
produce duplicates, because we weren't careful not to create multiple
REs with the same combination of owner, type, and title.

Now, be more careful. Disambiguate by adding an incremented numeric ID
at the end of each duplicate except the first one encounted.