If an RE's title is changed, re-index vocabs having it as publisher
The API method to update an RE didn't pay attention to the Solr index.

The Solr index includes titles of publishers.

So if an RE's title is changed, we need to re-index all vocabularies
that have it as a publisher.

Add a note to DeleteRelatedEntities explaining why nothing needs to be
done in _that_ class.