When deleting an RE, also check if it's used in any draft instances
Oops, we can get a database inconsistency if we allow deleting a
related entity, if there are any draft instances that use it.

We only checked for uses of the RE in current vocabulary instances.

Now, check _both_ current _and_ draft instances of all vocabularies
when determining if we will allow the deletion of an RE.