Make index on (vocabulary_id, end_date) not unique
Sigh, creating the index on the vocabularies table on the combination
(vocabulary_id, end_date) as a unique index currently fails, because
of completely failing to take into account the fact that Hibernate
reorders instructions.

So, e.g., when updating a current vocabulary, you get a constraint
violation because VocabularyModel's update is reordered to come after
its insert.

So, for now, don't make the index unique. We may revisit.  The code
would have to have em().flush() statements added in various places.