Move registry event generation into the model subsystem
Don't create registry events in the code for API methods, but in the
model subsystem instead.

In the specific case of updating related entities, don't register an
update to the RE unless something has actually changed.

As a result of moving the registry event generation into the model,
the test data for the model tests now has to include registry events.