Upgrade Checkstyle to release 8.10
Having upgraded my Eclipse to Oxygen.3A, Checkstyle is now run by
Eclipse using a Java 9 JVM, and now complains about imports of
java.sql in test code. I upgraded the Eclipse Checkstyle plugin to see
if it fixes the problem. It doesn't, but now I can report the problem.

Upgrade Checkstyle to release 8.10 here, as it seems to be OK in our
Ant builds.