Support special owner name "*" for subscriptions
Support subscribing to notifications for all vocabularies, by
introducing the special owner name "*", mapped internally by the owner
cache to the owner Id 0.

To subscribe to notifications for all vocabularies, subscribe to the
owner '*". Subscription/unsubscription from this special owner does
not (as yet) have any effect on any subscriptions to "real" owners.