Add RolesUtils.getOrgRolesByRoleId() method
Define a RolesUtils.getOrgRolesByRoleId() method that fetches cut-down
instances of organisational roles.  This special treatment for
organisational roles is needed, as they don't have
AuthenticationServiceId data, and so fetching entire rows fails.

This method will be used to get full names of organisational roles,
for inclusion in generated notification emails.