In subscriber tokens, use _ instead of , to separate components
We're going to use subscriber tokens in URL paths for the portal, and
it turns out that CodeIgniter has a (default) setting that rejects
commas in path components.

I prefer in this case to switch our generated tokens to use
underscores instead of commas.

So tokens now look like "2_abc" instead of "2,abc".