Add many classes for subscriptions and notifications
Add fixed-time model subsystem.

Add classes with utility methods for comparing the fields of two
entity instances and collecting the differences.

Add API method to get subscriptions for a subscriber.

Add classes for collecting the events for a notification period.

Add a class with a main method to send email notifications.