Don't use a transaction for GetSubscriptions
GetSubscriptions started a transaction, but never rolled it back.
This causes a leak of database connections which leads to eventual
failure of the webapp (sigh).

(The code was copy/pasted from PutSubscriptions, which uses a
transaction.)

It was wrong to use a transaction for this API method, so don't use
one.