Fix handling of vocab created and deleted during reporting period
When collecting a registry event for a vocabulary, we start by
creating a VocabularyDifferences instance for it and putting it in the
vocabularyIdMap Map.

But we need to take care to remove it from the Map if we find that the
vocabulary did not exist at both the start and end date/times of the
reporting period. Nothing will be reported for that vocabulary.