Improve comments in VocabularyDifferences and VersionDifferences
Change the class comment for VocabularyDifferences, and the
description of the finalResult field of both VocabularyDifferences and
VersionDifferences, to reflect how the implementation actually turned
out. I.e., the values of these fields is not so much based on the last
RegistryEvent recorded for the entity (though it should match!), but a
reflection of the final state in comparison with its state at the
start of the reporting period.