Improve package-info.java files and Javadoc overview; move RolesUtils
Go through all the package-info.java files and make the first sentence
of each Javadoc comment helpful in isolation, as it is included in the
Javadoc overview page.

Add an overview.html file, with some initial content to be included in
the Java overview page.

Group packages into top-level groups for the overview page.

As part of this work, it became clear that the RolesUtils class was in
the wrong package.  So, move the RolesUtils class to the correct
package.

It also became clear that there are other cases where classes should
be moved into a different package (and in one case, renamed). But
those other classes have been left where they are, and a comment
added.