Upgrade Commons Lang to 3.7; add Commons Text
I've been trying some experiments with the ReflectionDiffBuilder class
in Commons Lang 3.7. So bite the bullet and upgrade Commons Lang from
version 3.4 to 3.7.

There are now deprecations for the classes in the package
org.apache.commons.lang3; the classes have been moved to Commons
Text. So add Commons Text, and update our code to use it.