Strip leading/trailing whitespace from emails for subscription methods
Explicitly strip leading/trailing whitespace from email addresses
provided to subscription API methods.

Tweak the debug logging of the email address at the beginning of
createEmailSubscriptionVocabulary(): there was an extra space in the
log content, which suggested there might be a leading space in the
email address! But no. Remove the bogus space to make the log content
consistent across the three API methods.