Log error when migrating resource_owner_hosts to warn about time zone
In MigrateToolkitToRegistry, failed to pay attention to time zone
badness when migrating the contents of the resource_owner_hosts table,
which already had start_date and end_date columns.

Oh well, I don't want to spend much time on this any more, so log an
error indicating that a manual patch-up may be required.