Remove typo; include short name for owner when full name unknown
The Portal's "Manage Subscriptions" page showed "Unkknown owner name"
where the owner is not in the roles database. Not only is that a typo,
but we should do better, since we do currently have that situation for
ANDS-Curated.

SubscriptionDbSchemaMapperDecorator had a typo "Unkknown"; take the
opportunity to fix it better, by including the owner short name in the
value returned in the case where the owner's full name is not known
(i.e., is not in the roles database).