Email sending script: take daylight-savings switch into account
Sigh, another daylight-savings stuff-up.

When sending email notifications, take more core about when "midnight"
7 days ago is.

Instead of going back seven days _after_ converting today midnight to
UTC, do the arithmetic _before_ converting to UTC.

That means that for "7 days ago" we get midnight with the
daylight-savings setting that was in play 7 days ago, which might be
different from the setting that is in play "today".

The old code always computed a 168-hour reporting period. This was
wrong when the reporting period straddles a daylight-savings
changeover: in that case, "7 days" does _not_ equal 168 hours.

The new code does the arithmetic on the midnight in the local
timezone, before converting the result to UTC.