Remove @Produces annotation from lookupIRI API method
The lookupIRI API method had a @Produces annotation, that declared
that results were limited to plain text. As a result, callers that
requested anything else (e.g., XML) using an Accept request header got
a 406 response.

Just remove the @Produces annotation, so that the method gets to see
all requests, irrespective of the value of any Accept request header.

If the method then sends back a redirect, it will (and should) be the
"problem" of the new target URL to inspect an Accept header.