CC-2190 CC-2215 Change type of dcterms
In conf/ANDS-ELDAConfig-template.ttl.sample, change the defined type
of dcterms:created and dcterms:modified to xsd:dateTime, so that
filters using comparisons work correctly.

Define a conceptSearchEndPoint parameter that points to the /concept
endpoint, to use to make reliable links to filtered searches,
specifically, so that they work on the /resource endpoint.