CC-2387 Fix regression due to previous update of commons-lang3
The update of commons-lang3 from version 3.4 to 3.7 in commit
a3c2ea6883cc8c3da9d3264deb98d836d88cd71e had an unexpected
side-effect: it made it impossible to add more than one new
user-specified access point to the same version in one step.

The cause: commons-lang3 changed the behaviour of the reflection-based
StringBuilder so that it no longer accepted null parameters.

The fix: be more careful about the values of parameters passed to that
StringBuilder.

Add a set of regression tests for the various cases.

Of note: oops, I discovered some inconsistency in the handling of
user-specified access points of type sesameDownload.

The validation process lets them through, but the model doesn't
support them. This is something to fix later.