search API method: add comment explaining form/query parameters
I didn't realise that Jersey has a feature, enabled by default, that
allows a client to provide a parameter value as a query parameter,
where the parameter has been declared in the server-side code with the
@FormParam annotation.

Document that feature where it applies, i.e., for the search API
method.

Note that because of the use of our custom @FormContentTypeFilter
annotation, the client still has to specify the Content-Type as
"application/x-www-form-urlencoded", but you don't have to provide a
body!