CC-2444: Support either integer or string "pp" filter for search
Allow the API caller to specify a search filter with a value for "pp"
that is either an integer or a string.

Previously, if the "pp" filter was specified as a string, an
exception leaked out through Tomcat to the client as a
stacktrace.

See commit a7f35f7a for the earlier work to support this for the "p"
filter.