CC-2447: updateVocabulary: check owner of updated vocabulary
The updateVocabulary method didn't check that the updated vocabulary's
owner was one of the user's allowed values. As long as you owned a
vocabulary, you could change the owner to _any value_.

Now, check the updated vocabulary's owner value to make sure that the
user has permission to create/update vocabularies with that owner
value.