Change to empty @Produces annotation on download API methods
The download API methods had @Produces annotations, that declared that
results were limited to "application/octet-stream". As a result,
callers that requested anything else (e.g., XML) using an Accept
request header got a 406 response.

Removing the @Produces annotations "should" be enough, but Swagger UI
then uses the default settings defined in SwaggerInterface, i.e.,
either XML or JSON. Removing the parameter from the @Produces
annotations has the desired effect, and means that Swagger UI offers
"*/* as the single option in the "Response Content Type" dropdown.

So remove the parameter from the @Produces annotation, so that the
download methods get to see all requests, irrespective of the value of
any Accept request header.