Include version Id in error message from VersionsModel
In VersionsModel, for the error "Attempt to update version that does
not belong to this vocabulary", include the version Id in the error
message.

This error message is generated in two places; in the code for drafts,
the error message already did include the version Id, but this was not
the case for updating the current version. Include the version Id in
the error message for updating the current version.