CC-1488 Refactor Analytics and add many more fields
The Analytics class was a mess; it had two inconsistent copies of much
code. Remove duplication and rationalize.

Define more logging fields. Define top-level maps/groups of fields, so
that there is now a hierarchy of the log fields to match the portal
logs more closely.

Make the invocations of analytics logging more consistent, and do more
logging (especially, of errors).