CC-1488: Analytics "vocabulary.entity_status" -> "vocabulary.status"
When all the analytics log fields were at the top level, it made sense
to have a field name "entity_status" that did not have the potential
ambiguity of just "status".

But now that this field is embedded within a top-level field
"vocabulary", we can now use just "status".

So, rename the field.