CC-1488 Refactor getVocabularyById for analytics for getV...ByIdEdit
The getVocabularyByIdEdit method delegated to getVocabularyById for
published/deprecated instances. But that meant that an invocation of
getVocabularyByIdEdit also did the analytics logging that
getVocabularyById did. But we want getVocabularyByIdEdit to log the
user profile.

So refactor the core of getVocabularyById into a helper method
getVocabularyByIdHelper that doesn't do any analytics logging.

Invoke getVocabularyByIdHelper in both getVocabularyById and
getVocabularyByIdEdit, but do different analytics logging.