CC-1488 Analytics logging for get owned vocab, get/put/delete REs
When doing analytics logging of methods that create/read/update/delete
REs, group the RE-related data into an object.

Use "id" also as the field name for RE Ids. NB: because of the way
updateMarkerWithAdditionalFields() is coded, this will work up till a
point where we need a log entry that contains _both_ a vocab and an
RE.

Add an analytics log entry for the API method to get owned
vocabularies.