CC-1488 Add owner and slug fields to result of getOwnedVocabularies
In order to be able to do portal-side analytics logging of vocabulary
deletion in a way that's consistent with the analytics logging for
vocabulary creation and updates, the My Vocabs portal page needs
vocabulary metadata not currently provided by the getOwnedVocabularies
method: the vocabulary owner and (title) slug.

So add owner and slug to the attributes of the Registry Schema
owned-vocabulary element, and have getOwnedVocabularies fill them in.