Make Subtask.compareTo() "consistent with equals"
The implementation of Subtask.compareTo() failed with an exception, if
the Subtasks being compared had the same priority, provider type,
operation, and provider, and both had subtask properties.

(This currently never happens in practice, because we never try to add
two basically-identical Subtasks that both have subtask properties to
the same Task. But this could happen in future.)

This is because the comparison of subtask properties was done in a way
that assumed that Map implements Comparable. But it doesn't.

So convert each Subtask's subtask properties into an array of Strings
that _can_ be compared.

So we have now made the compareTo() method "consistent with equals" in
the sense defined by the Comparable interface.